der: move `PemWriter` to its own module
Internal refactor in preparation for adding `PemReader`.
